Re-export old dances
When the old DDR dances were originally exported, I retargetted the arm
bone to a _clavicle_ bone on the new skeleton by mistake. This meant the
dancers would move their _clavicles_ in ways that were unexpected, and
frankly just looked bad.

With this change, the dances were retargeted appropriately, and things
look a lot nicer